Tests for current and legacy parser
