+++ src/test/java/ua_parser/ParserTestBase.java
@@ -33,15 +33,11 @@ import org.yaml.snakeyaml.Yaml;
  *
  * @author Steve Jiang (@sjiang) <gh at iamsteve com>
  */
-public class ParserTest {
-  final String TEST_RESOURCE_PATH = "/ua_parser/legacy/";
+abstract public class ParserTestBase {
   Yaml yaml = new Yaml();
   Parser parser;
 
-  @Before
-  public void initParser() throws Exception {
-    parser = new Parser();
-  }
+  abstract String getResourcePath();
 
   @Test
   public void testParseUserAgent() {
@@ -75,22 +71,6 @@ public class ParserTest {
   }
 
   @Test
-  public void testParseAll() {
-    String agentString1 = "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.4; fr; rv:1.9.1.5) Gecko/20091102 Firefox/3.5.5,gzip(gfe),gzip(gfe)";
-    String agentString2 = "Mozilla/5.0 (iPhone; CPU iPhone OS 5_1_1 like Mac OS X) AppleWebKit/534.46 (KHTML, like Gecko) Version/5.1 Mobile/9B206 Safari/7534.48.3";
-
-    Client expected1 = new Client(new UserAgent("Firefox", "3", "5", "5"),
-                                  new OS("Mac OS X", "10", "4", null, null),
-                                  new Device("Other", null, null));
-    Client expected2 = new Client(new UserAgent("Mobile Safari", "5", "1", null),
-                                  new OS("iOS", "5", "1", "1", null),
-                                  new Device("iPhone", "Apple", "iPhone"));
-
-    assertThat(parser.parse(agentString1), is(expected1));
-    assertThat(parser.parse(agentString2), is(expected2));
-  }
-
-  @Test
   public void testReplacementQuoting() throws Exception {
     String testConfig = "user_agent_parsers:\n"
                       + "  - regex: 'ABC([\\\\0-9]+)'\n"
@@ -115,7 +95,7 @@ public class ParserTest {
   }
 
   void testUserAgentFromYaml(String filename) {
-    InputStream yamlStream = this.getClass().getResourceAsStream(TEST_RESOURCE_PATH + filename);
+    InputStream yamlStream = this.getClass().getResourceAsStream(getResourcePath() + filename);
 
     @SuppressWarnings("unchecked")
     Map<String, List<Map<String,String>>> entries = (Map<String, List<Map<String,String>>>)yaml.load(yamlStream);
@@ -133,7 +113,7 @@ public class ParserTest {
   }
 
   void testOSFromYaml(String filename) {
-    InputStream yamlStream = this.getClass().getResourceAsStream(TEST_RESOURCE_PATH + filename);
+    InputStream yamlStream = this.getClass().getResourceAsStream(getResourcePath() + filename);
 
     @SuppressWarnings("unchecked")
     Map<String, List<Map<String,String>>> entries = (Map<String, List<Map<String,String>>>)yaml.load(yamlStream);
@@ -149,7 +129,7 @@ public class ParserTest {
   }
 
   void testDeviceFromYaml(String filename) {
-    InputStream yamlStream = this.getClass().getResourceAsStream(TEST_RESOURCE_PATH + filename);
+    InputStream yamlStream = this.getClass().getResourceAsStream(getResourcePath() + filename);
 
     @SuppressWarnings("unchecked")
     Map<String, List<Map<String,String>>> entries = (Map<String, List<Map<String,String>>>)yaml.load(yamlStream);
@@ -162,8 +142,5 @@ public class ParserTest {
     }
   }
 
-  Parser parserFromStringConfig(String configYamlAsString) throws Exception {
-    InputStream yamlInput = new ByteArrayInputStream(configYamlAsString.getBytes("UTF8"));
-    return new Parser(yamlInput);
-  }
+  abstract Parser parserFromStringConfig(String configYamlAsString) throws Exception;
 }
